Test origanization/cleanup.
