+++ core/src/test/scala/org/locationtech/rasterframes/TestData.scala
@@ -180,7 +180,9 @@ trait TestData {
     TestData.randomTile(cols, rows, UByteConstantNoDataCellType)
   )).map(ProjectedRasterTile(_, extent, crs)) :+ null
 
-  def lazyPRT = RasterRef(RFRasterSource(TestData.l8samplePath), 0, None, None).tile
+  def rasterRef = RasterRef(RFRasterSource(TestData.l8samplePath), 0, None, None)
+  def lazyPRT = rasterRef.tile
+
 
   object GeomData {
     val fact = new GeometryFactory()
