+++ core/src/test/scala/org/locationtech/rasterframes/functions/TileFunctionsSpec.scala
@@ -23,13 +23,12 @@ package org.locationtech.rasterframes.functions
 import java.io.ByteArrayInputStream
 
 import geotrellis.proj4.CRS
-import geotrellis.raster.testkit.RasterMatchers
 import geotrellis.raster._
+import geotrellis.raster.testkit.RasterMatchers
 import javax.imageio.ImageIO
 import org.apache.spark.sql.Encoders
 import org.apache.spark.sql.functions.sum
 import org.locationtech.rasterframes._
-import org.locationtech.rasterframes.stats._
 import org.locationtech.rasterframes.tiles.ProjectedRasterTile
 import org.locationtech.rasterframes.util.ColorRampNames
 
@@ -89,6 +88,7 @@ class TileFunctionsSpec extends TestEnvironment with RasterMatchers {
 
       checkDocs("rf_convert_cell_type")
     }
+
     it("should change NoData value") {
       val df = Seq((TestData.injectND(7)(three), TestData.injectND(12)(two))).toDF("three", "two")
 
@@ -107,21 +107,19 @@ class TileFunctionsSpec extends TestEnvironment with RasterMatchers {
       // Should maintain original cell type.
       ndCT.select(rf_cell_type($"two")).first().withDefaultNoData() should be(ct.withDefaultNoData())
     }
+
     it("should interpret cell values with a specified cell type") {
       checkDocs("rf_interpret_cell_type_as")
-      val df = Seq(randNDPRT)
-        .toDF("t")
+      val df = Seq(randNDPRT).toDF("t")
         .withColumn("tile", rf_interpret_cell_type_as($"t", "int8raw"))
       val resultTile = df.select("tile").as[Tile].first()
 
       resultTile.cellType should be(CellType.fromName("int8raw"))
       // should have same number of values that are -2 the old ND
-      val countOldNd = df
-        .select(
-          rf_tile_sum(rf_local_equal($"tile", ct.noDataValue)),
-          rf_no_data_cells($"t")
-        )
-        .first()
+      val countOldNd = df.select(
+        rf_tile_sum(rf_local_equal($"tile", ct.noDataValue)),
+        rf_no_data_cells($"t")
+      ).first()
       countOldNd._1 should be(countOldNd._2)
 
       // should not have no data any more (raw type)
@@ -239,158 +237,16 @@ class TileFunctionsSpec extends TestEnvironment with RasterMatchers {
     }
 
     it("should get the CRS of a RasterRef") {
-      fail()
-      //val e = Seq(Tuple1(TestData.rasterRef)).toDF("ref").select(rf_crs($"ref")).first()
-      //e should be(rasterRef.crs)
+      val e = Seq(Tuple1(TestData.rasterRef)).toDF("ref").select(rf_crs($"ref")).first()
+      e should be(rasterRef.crs)
     }
 
     it("should get the Extent of a RasterRef") {
-      fail()
-      //val e = Seq(Tuple1(rasterRef)).toDF("ref").select(rf_extent($"ref")).first()
-      //e should be(rasterRef.extent)
+      val e = Seq(Tuple1(rasterRef)).toDF("ref").select(rf_extent($"ref")).first()
+      e should be(rasterRef.extent)
     }
   }
-  describe("per-tile stats") {
-    it("should compute data cell counts") {
-      val df = Seq(TestData.injectND(numND)(two)).toDF("two")
-      df.select(rf_data_cells($"two")).first() shouldBe (cols * rows - numND).toLong
 
-      val df2 = randNDTilesWithNull.toDF("tile")
-      df2
-        .select(rf_data_cells($"tile") as "cells")
-        .agg(sum("cells"))
-        .as[Long]
-        .first() should be(expectedRandData)
-
-      checkDocs("rf_data_cells")
-    }
-    it("should compute no-data cell counts") {
-      val df = Seq(TestData.injectND(numND)(two)).toDF("two")
-      df.select(rf_no_data_cells($"two")).first() should be(numND)
-
-      val df2 = randNDTilesWithNull.toDF("tile")
-      df2
-        .select(rf_no_data_cells($"tile") as "cells")
-        .agg(sum("cells"))
-        .as[Long]
-        .first() should be(expectedRandNoData)
-
-      checkDocs("rf_no_data_cells")
-    }
-
-    it("should properly count data and nodata cells on constant tiles") {
-      val rf = Seq(randPRT).toDF("tile")
-
-      val df = rf
-        .withColumn("make", rf_make_constant_tile(99, 3, 4, ByteConstantNoDataCellType))
-        .withColumn("make2", rf_with_no_data($"make", 99))
-
-      val counts = df
-        .select(
-          rf_no_data_cells($"make").alias("nodata1"),
-          rf_data_cells($"make").alias("data1"),
-          rf_no_data_cells($"make2").alias("nodata2"),
-          rf_data_cells($"make2").alias("data2")
-        )
-        .as[(Long, Long, Long, Long)]
-        .first()
-
-      counts should be((0l, 12l, 12l, 0l))
-    }
-
-    it("should detect no-data tiles") {
-      val df = Seq(nd).toDF("nd")
-      df.select(rf_is_no_data_tile($"nd")).first() should be(true)
-      val df2 = Seq(two).toDF("not_nd")
-      df2.select(rf_is_no_data_tile($"not_nd")).first() should be(false)
-      checkDocs("rf_is_no_data_tile")
-    }
-
-    it("should evaluate exists and for_all") {
-      val df0 = Seq(zero).toDF("tile")
-      df0.select(rf_exists($"tile")).first() should be(false)
-      df0.select(rf_for_all($"tile")).first() should be(false)
-
-      Seq(one).toDF("tile").select(rf_exists($"tile")).first() should be(true)
-      Seq(one).toDF("tile").select(rf_for_all($"tile")).first() should be(true)
-
-      val dfNd = Seq(TestData.injectND(1)(one)).toDF("tile")
-      dfNd.select(rf_exists($"tile")).first() should be(true)
-      dfNd.select(rf_for_all($"tile")).first() should be(false)
-
-      checkDocs("rf_exists")
-      checkDocs("rf_for_all")
-    }
-    it("should find the minimum cell value") {
-      val min = randNDPRT.toArray().filter(c => isData(c)).min.toDouble
-      val df = Seq(randNDPRT).toDF("rand")
-      df.select(rf_tile_min($"rand")).first() should be(min)
-      df.selectExpr("rf_tile_min(rand)").as[Double].first() should be(min)
-      checkDocs("rf_tile_min")
-    }
-
-    it("should find the maximum cell value") {
-      val max = randNDPRT.toArray().filter(c => isData(c)).max.toDouble
-      val df = Seq(randNDPRT).toDF("rand")
-      df.select(rf_tile_max($"rand")).first() should be(max)
-      df.selectExpr("rf_tile_max(rand)").as[Double].first() should be(max)
-      checkDocs("rf_tile_max")
-    }
-    it("should compute the tile mean cell value") {
-      val values = randNDPRT.toArray().filter(c => isData(c))
-      val mean = values.sum.toDouble / values.length
-      val df = Seq(randNDPRT).toDF("rand")
-      df.select(rf_tile_mean($"rand")).first() should be(mean)
-      df.selectExpr("rf_tile_mean(rand)").as[Double].first() should be(mean)
-      checkDocs("rf_tile_mean")
-    }
-
-    it("should compute the tile summary statistics") {
-      val values = randNDPRT.toArray().filter(c => isData(c))
-      val mean = values.sum.toDouble / values.length
-      val df = Seq(randNDPRT).toDF("rand")
-      val stats = df.select(rf_tile_stats($"rand")).first()
-      stats.mean should be(mean +- 0.00001)
-
-      val stats2 = df
-        .selectExpr("rf_tile_stats(rand) as stats")
-        .select($"stats".as[CellStatistics])
-        .first()
-      stats2 should be(stats)
-
-      df.select(rf_tile_stats($"rand") as "stats")
-        .select($"stats.mean")
-        .as[Double]
-        .first() should be(mean +- 0.00001)
-      df.selectExpr("rf_tile_stats(rand) as stats")
-        .select($"stats.no_data_cells")
-        .as[Long]
-        .first() should be <= (cols * rows - numND).toLong
-
-      val df2 = randNDTilesWithNull.toDF("tile")
-      df2
-        .select(rf_tile_stats($"tile")("data_cells") as "cells")
-        .agg(sum("cells"))
-        .as[Long]
-        .first() should be(expectedRandData)
-
-      checkDocs("rf_tile_stats")
-    }
-
-    it("should compute the tile histogram") {
-      val df = Seq(randNDPRT).toDF("rand")
-      val h1 = df.select(rf_tile_histogram($"rand")).first()
-
-      val h2 = df
-        .selectExpr("rf_tile_histogram(rand) as hist")
-        .select($"hist".as[CellHistogram])
-        .first()
-
-      h1 should be(h2)
-
-      checkDocs("rf_tile_histogram")
-    }
-  }
 
   describe("conversion operations") {
     it("should convert tile into array") {
